+++ src/civilization/game_phases/plugin.rs
@@ -1,0 +1,22 @@
+use crate::GameState;
+use bevy::app::Update;
+use bevy::prelude::{in_state, AppExtStates, IntoSystemConfigs, Plugin};
+use crate::civilization::game_phases::structs::GameActivity;
+use crate::civilization::game_phases::systems::{direct_game_phases, print_names_of_phases};
+
+pub struct GamePhasesPlugin;
+
+impl Plugin for GamePhasesPlugin {
+    fn build(&self, app: &mut bevy::app::App) {
+        app
+            .add_sub_state::<GameActivity>()
+            .add_systems(
+                Update, (
+                    print_names_of_phases.run_if(in_state(GameState::Playing)),
+                    direct_game_phases.run_if(in_state(GameState::Playing)),
+                ),
+            )
+        ;
+    }
+}
+
