+++ src/civilization/general/plugin.rs
@@ -3,7 +3,7 @@ use bevy::prelude::{in_state, Component, Entity, Event, IntoSystemConfigs, OnEnt
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 use crate::civilization::census::plugin::CensusPlugin;
 use crate::civilization::census::resources::GameInfoAndStuff;
-use crate::civilization::game_phases::GamePhasesPlugin;
+use crate::civilization::game_phases_plugin::GamePhasesPlugin;
 use crate::civilization::general::systems::{connect_areas, move_tokens_from_stock_to_area, setup_game};
 use crate::civilization::movement::MovementPlugin;
 use crate::civilization::population_expansion::PopulationExpansionPlugin;
