+++ src/civilization/mod.rs
@@ -1,5 +1,6 @@
 pub mod population_expansion;
 pub mod movement;
-pub mod game_phases;
+pub mod game_phases_plugin;
 pub mod census;
 pub mod general;
+pub mod game_phases;
